Bugs fixed, code sorted, tests written and passed
