+++ test_scripts/maths_test.ncs
@@ -1,0 +1,23 @@
+#A linear test for maths in node call
+
+main>
+    assign(foo, 4);
+    assign(bar, 7);
+    assign(additionResult, add(foo, bar));
+    assign(subtractionResult, sub(bar, foo));
+    assign(multiplicationResult, mul(foo, bar));
+    assign(multiplicationResultCommutative, mul(bar, foo));
+    assign(divisionResult, div(bar, foo));
+    assign(divisionResultFloat, div(float(bar), float(foo)));
+
+    #Right now, you cannot print non string types. Will get round to
+    #adding this again one day
+    assign(fooS, string(foo));
+    assign(barS, string(bar));
+    println("Addition of ", fooS, " and ", barS, ":\t", string(additionResult));
+    println("Subtraction of ", barS, " with ", fooS, ":\t", string(subtractionResult));
+    println("Multiplication of ", fooS, " and ", barS, ":\t", string(multiplicationResult));
+    println("And multiplication of ", barS, " and ", fooS, ":\t", string(multiplicationResultCommutative));
+    println("Division of ", barS, " with ", fooS, ":\t", string(divisionResult));
+    println("Division of ", string(float(bar)), " and ", string(float(foo)), ":\t", string(divisionResultFloat));
+>
